Make error handling consistent for all font types
Some fonts respond with empty strings when characters cannot be decoded.
Others respond with an error. A user of the library cannot really do much with
an error, but expects the contents to be extracted on a best-effort basis.
Other imprecisions already introduce defects into the text, such as footers that break
the linear textflow, so simply deleting the characters appears a reasonable solution.